src/api/pyproject.toml
[tool.poetry]
name = "picopix-api"
version = "0.1.0"
description = ""
authors = [
    "Mohamed CHELALI <mouha.chelali@gmail.com>",
    "Yannick OREAL <axee@laposte.net>",
    "Daniel LEWANDOWSKI <lewandowski_daniel_s@yahoo.fr>",
]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9"
python-dotenv = "^1.0.1"
python-multipart = "^0.0.17"
pyjwt = "^2.10.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
psycopg2-binary = "^2.9.10"
uvicorn = "^0.32.0"
fastapi = "^0.115.5"
sqlalchemy = "^2.0.36"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
